rename routes dir, add cron config
